+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wxtm-bridge-frontend",
-  "version": "0.1.5",
+  "version": "0.1.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -23,7 +23,8 @@
     "react-hook-form": "^7.56.2",
     "react-icons": "^5.5.0",
     "viem": "2.x",
-    "wagmi": "^2.14.16"
+    "wagmi": "^2.14.16",
+    "zustand": "^5.0.4"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
