+++ types/tapplet/account.ts
@@ -1,0 +1,4 @@
+export interface AccountData {
+  account_id: number
+  address: string
+}
