+++ types/tapplet/index.ts
@@ -1,2 +1,2 @@
-export * from "./signer";
-export * from "./types";
+export * from './account'
+export * from './signer'
